+++ pyqueen/io/image.py
